+++ CHANGES.md
@@ -51,6 +51,7 @@
 
 ## Highlights
 
+* Twister2 Runner ([BEAM-7304](https://issues.apache.org/jira/browse/BEAM-7304))
 * New highly anticipated feature X added to Python SDK ([BEAM-X](https://issues.apache.org/jira/browse/BEAM-X)).
 * New highly anticipated feature Y added to Java SDK ([BEAM-Y](https://issues.apache.org/jira/browse/BEAM-Y)).